Update Compose Multiplatform to 1.9.2

+++ library/src/commonMain/kotlin/ir/mahozad/multiplatform/wavyslider/material3/WavySlider.kt
@@ -675,7 +675,7 @@ private fun WavySliderImpl(
                     .wrapContentWidth()
                     .onSizeChanged {
                         @Suppress("INVISIBLE_REFERENCE")
-                        state.thumbWidth = it.width.toFloat()
+                        state.thumbWidth = it.width
                     }
             ) {
                 thumb(state)
